Change cave_on_path to cave_next_on_path
(Same for passages) 
Saw some weird behaviour where an agent might claim the cave twice if it's starts in the cave, and then it will block the cave/passage until it returns.

Now find_wanted_passages_and_cave should only find the passage/cave directly connected to the next move. 
This was a problem in e.g. alpha one where others might want the passage while it goes and gets the next box.